strands-langfuse/requirements.txt
langfuse>=2.53.3
strands-agents[otel]>=0.2.0
strands-agents-tools>=0.1.8
python-dotenv>=1.0.0
requests>=2.31.0
boto3>=1.34.0